+++ modules/canvaskit/npm_build/types/canvaskit-wasm-tests.ts
@@ -617,6 +617,14 @@ function paragraphBuilderTests(CK: CanvasKit, fontMgr?: FontMgr, paint?: Paint) 
     builder.setGraphemeBreaksUtf16(new Uint32Array(10));
     builder.setLineBreaksUtf16(new Uint32Array(10));
     const paragraph3 = builder.build(); // $ExpectType Paragraph
+
+    const fontCollection = CK.FontCollection.Make(); // $ExpectType FontCollection
+    fontCollection.enableFontFallback();
+    fontCollection.setDefaultFontManager(fontSrc);
+    const fcBuilder = CK.ParagraphBuilder.MakeFromFontCollection(// $ExpectType ParagraphBuilder
+        paraStyle, fontCollection);
+    fcBuilder.addText('12345');
+    const fcParagraph = fcBuilder.build();
 }
 
 function pathEffectTests(CK: CanvasKit, path?: Path) {